Unify navigation into single GlobalNav with 4 glassmorphism variants
Replace separate HomepageNav and Header/GlobalNav with a unified
GlobalNav component using the glass button variant. Supports 4
variations based on auth state and page context (default/timeline).
