--- src/components/Homepage/HomepageNav.tsx
@@ -1,96 +1,0 @@
-import { useState } from 'react';
-import { LayoutList, Video } from 'lucide-react';
-import { useAuth } from '../../contexts/AuthContext';
-import { supabase } from '../../lib/supabase';
-import { Button } from '@/components/ui/button';
-import {
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-} from '@/components/ui/dialog';
-import { FeedbackPanel } from '@/components/FeedbackPanel/FeedbackPanel';
-
-export function HomepageNav() {
-  const { user } = useAuth();
-  const [isFeedbackOpen, setIsFeedbackOpen] = useState(false);
-  const [isVideoTutorialOpen, setIsVideoTutorialOpen] = useState(false);
-
-  const handleSignOut = async () => {
-    try {
-      await supabase.auth.signOut();
-    } catch (error) {
-      console.error('Error signing out:', error);
-    }
-  };
-
-  return (
-    <div className="bg-surface-primary">
-      <div className="mx-auto px-6 py-3 flex justify-between items-center">
-        <div className="flex items-center gap-2">
-          <LayoutList size={20} className="text-text-tertiary" />
-          {user && (
-            <span className="text-sm font-medium text-text-primary">Timelines</span>
-          )}
-        </div>
-
-        <div className="flex items-center gap-2">
-          <Button
-            variant="outline"
-            size="sm"
-            onClick={() => setIsVideoTutorialOpen(true)}
-          >
-            How It Works
-          </Button>
-          {user && (
-            <>
-              <Button
-                variant="outline"
-                size="sm"
-                onClick={() => setIsFeedbackOpen(true)}
-              >
-                Feedback
-              </Button>
-              <Button
-                variant="outline"
-                size="sm"
-                onClick={handleSignOut}
-              >
-                Log Out
-              </Button>
-            </>
-          )}
-        </div>
-      </div>
-
-      <FeedbackPanel open={isFeedbackOpen} onOpenChange={setIsFeedbackOpen} />
-
-      <Dialog open={isVideoTutorialOpen} onOpenChange={setIsVideoTutorialOpen}>
-        <DialogContent className="max-w-[550px]">
-          <DialogHeader>
-            <DialogTitle>Quick Tutorial to Get Started</DialogTitle>
-          </DialogHeader>
-          <div className="space-y-4">
-            <p className="text-muted-foreground">
-              This video tutorial walks through how to start building your timeline by adding events, editing categories, customizing timeline settings, and importing or exporting data to build faster.
-            </p>
-            <div className="aspect-video">
-              <a
-                href="https://www.loom.com/share/f19575818a9341d4a266c482af981ba2"
-                target="_blank"
-                rel="noopener noreferrer"
-                className="block w-full h-full bg-secondary rounded-lg flex items-center justify-center hover:bg-secondary/80 transition-colors"
-              >
-                <div className="text-center p-6">
-                  <Video size={48} className="mx-auto mb-4" />
-                  <p>Click to watch the tutorial video on Loom</p>
-                  <p className="text-sm text-muted-foreground mt-2">The video will open in a new tab</p>
-                </div>
-              </a>
-            </div>
-          </div>
-        </DialogContent>
-      </Dialog>
-    </div>
-  );
-}
